Moves DataModule to train.py
